refactor(job): strip historical/design-doc framing from comments; decompose execution_hooks; fold ExecutionScheduled into NewJob's initial event batch
Comment cleanup pass across the whole PR diff (parked rows + head-swap
+ commit-hook decomposition): removed the large majority of inline `//`
comments (redundant with doc comments, or narrating design history --
handoff addendum section numbers, PR/commit/CI-run citations, "bugbot"/
"automated review"/"revert-to-red" attribution, references to the
superseded born-claimed design). Doc comments (`///`/`//!`) are kept
but rewritten to describe only the current code's behavior and the
still-true reasons for its shape -- hazard/invariant/race documentation
is preserved, historical narration around it is not. One real
stale-reference fix caught along the way: `JobInitializer::short_circuit()`'s
doc still described the old born-claimed trade-off (no fairness
guarantee); rewritten to describe the actual current head-swap
behavior (fairness is structural, nothing to opt out of for that
reason).

execution_hooks.rs decomposed into a module: execution_hooks/{mod,
insert,promote}.rs -- ExecutionInsertHook/NewExecutionRow in insert.rs,
PromoteHeadsHook/PromotedRow in promote.rs, mod.rs re-exporting both at
the same crate::execution_hooks:: path so no other file's imports
change.

NewJob gains a required `schedule_at` field; `IntoEvents` now emits
`ExecutionScheduled { attempt: 1, scheduled_at }` alongside
`Initialized` in ONE initial event batch (EntityEvents::init already
supports N initial events -- confirmed against es-entity's generated
create_in_op, which UNNESTs an arbitrary event array with no
single-event assumption). This removes a full create-then-update round
trip from every spawn path: create_job_internal, spawn_all_in_op,
KeyedJobSpawner::spawn, and ResidentJobSpawner::spawn all now pass
schedule_at into the builder and never call the now-dead
Job::schedule_execution (removed) or a second update_in_op. schedule_at
is required, not optional, because every real spawn path has it
available before construction; the handful of test-only NewJob
construction sites (repo.rs's jobs-table constraint tests, poller.rs's
raw-SQL reconciler fixtures) now pass a throwaway `chrono::Utc::now()`
rather than the type allowing a state no real caller should construct.

KeyedInsert and insert_keyed_execution moved from spawner.rs into
keyed.rs, next to KeyedJobSpawner -- spawner.rs no longer needs to know
about keyed-specific insert semantics. insert_keyed_execution's own
signature shrinks (no more `repo`/`&mut Job` params) now that it
doesn't need to append+persist ExecutionScheduled itself.

Verification: full 144-test suite green (one flaky-under-load test
excluded per established precedent, itself unaffected by this diff and
confirmed passing standalone); nix flake check clean; sqlx offline
cache regenerated and verified against a from-scratch SQLX_OFFLINE
build. Three apparent failures during iteration
(test_multi_day_scheduling_with_artificial_clock,
partial_claim_sleeps_until_next_due,
test_keep_alive_protects_live_own_instance_jobs) were diagnosed as
pre-existing dev-DB pollution from this session's own earlier test
runs (test_keep_alive's fixed, non-randomized "infinite-listener"
job_type had accumulated leftover running rows from prior runs that
never reached jobs.shutdown()) -- not a regression; all three pass
cleanly standalone and the full suite is green after a one-time DB
cleanup.
