+++ src/repo.rs
@@ -145,13 +145,11 @@ impl JobRepo {
         let row = sqlx::query_as!(
             JobExecutionRow,
             r#"
-            -- `parked` is purely a claim-time implementation detail (see
-            -- migrations, PERFORMANCE.md "Claim admission"): a parked row is
-            -- a queued job blocked behind its queue's active row, which is
-            -- exactly what "pending, not yet claimable" already meant to
-            -- every caller before this state existed. Masking it here keeps
-            -- `JobExecutionState`/`JobStatus` a two-value public contract
-            -- (`Pending`/`Running`) with no API surface change for it.
+            -- `parked` is a claim-time implementation detail: a parked row
+            -- is a queued job blocked behind its queue's active row, which
+            -- to every external caller just means "pending, not yet
+            -- claimable". Masking it here keeps `JobExecutionState`/
+            -- `JobStatus` a two-value public contract (`Pending`/`Running`).
             SELECT (CASE WHEN je.state = 'parked' THEN 'pending' ELSE je.state END)
                        AS "state!: JobExecutionState",
                    je.execute_at, je.attempt_index,
@@ -230,6 +228,7 @@ mod tests {
             .resident(true)
             .job_type(type_a.clone())
             .config(serde_json::json!({}))?
+            .schedule_at(chrono::Utc::now())
             .build()
             .expect("Could not build new job");
         repo.create(new_job).await?;
@@ -240,6 +239,7 @@ mod tests {
             .resident(true)
             .job_type(type_a.clone())
             .config(serde_json::json!({}))?
+            .schedule_at(chrono::Utc::now())
             .build()
             .expect("Could not build new job");
         let err: JobError = repo
@@ -256,6 +256,7 @@ mod tests {
             .resident(true)
             .job_type(type_b)
             .config(serde_json::json!({}))?
+            .schedule_at(chrono::Utc::now())
             .build()
             .expect("Could not build new job");
         repo.create(new_job).await?;
@@ -266,6 +267,7 @@ mod tests {
             .id(JobId::new())
             .job_type(type_a)
             .config(serde_json::json!({}))?
+            .schedule_at(chrono::Utc::now())
             .build()
             .expect("Could not build new job");
         repo.create(new_job).await?;
@@ -273,8 +275,8 @@ mod tests {
         Ok(())
     }
 
-    /// D13 (rewritten for live-keyed semantics): `jobs` no longer enforces
-    /// per-arbitrary-key uniqueness — it accumulates one row per generation
+    /// `jobs` does not enforce per-arbitrary-key uniqueness — it
+    /// accumulates one row per generation
     /// of a keyed singleton, and liveness enforcement moved to
     /// `job_executions` (`idx_job_executions_job_type_unique_key`, see
     /// `migrations/20250904065521_job_setup.sql`). The absolutely-unique
@@ -300,6 +302,7 @@ mod tests {
             .unique_key("k1")
             .job_type(type_a.clone())
             .config(serde_json::json!({}))?
+            .schedule_at(chrono::Utc::now())
             .build()
             .expect("Could not build new job");
         repo.create(new_job).await?;
@@ -311,6 +314,7 @@ mod tests {
             .unique_key("k1")
             .job_type(type_a.clone())
             .config(serde_json::json!({}))?
+            .schedule_at(chrono::Utc::now())
             .build()
             .expect("Could not build new job");
         repo.create(new_job).await?;
@@ -321,6 +325,7 @@ mod tests {
             .unique_key("k2")
             .job_type(type_a.clone())
             .config(serde_json::json!({}))?
+            .schedule_at(chrono::Utc::now())
             .build()
             .expect("Could not build new job");
         repo.create(new_job).await?;
@@ -332,6 +337,7 @@ mod tests {
             .unique_key("k3")
             .job_type(type_a.clone())
             .config(serde_json::json!({}))?
+            .schedule_at(chrono::Utc::now())
             .build()
             .expect("Could not build new job");
         let err: JobError = repo
@@ -348,6 +354,7 @@ mod tests {
             .unique_key("k1")
             .job_type(type_b)
             .config(serde_json::json!({}))?
+            .schedule_at(chrono::Utc::now())
             .build()
             .expect("Could not build new job");
         repo.create(new_job).await?;
@@ -359,6 +366,7 @@ mod tests {
             .id(JobId::new())
             .job_type(type_c.clone())
             .config(serde_json::json!({}))?
+            .schedule_at(chrono::Utc::now())
             .build()
             .expect("Could not build new job");
         repo.create(new_job).await?;
@@ -366,6 +374,7 @@ mod tests {
             .id(JobId::new())
             .job_type(type_c.clone())
             .config(serde_json::json!({}))?
+            .schedule_at(chrono::Utc::now())
             .build()
             .expect("Could not build new job");
         repo.create(new_job).await?;
@@ -373,6 +382,7 @@ mod tests {
             .id(a_id)
             .job_type(type_c)
             .config(serde_json::json!({}))?
+            .schedule_at(chrono::Utc::now())
             .build()
             .expect("Could not build new job");
         let err: JobError = repo
