+++ src/resident.rs
@@ -207,23 +207,22 @@ where
         fields(job_type = %self.job_type)
     )]
     pub async fn spawn(self, config: Config) -> Result<JobHandle, JobError> {
+        let mut op = self.repo.begin_op_with_clock(&self.clock).await?;
+        let schedule_at = op.maybe_now().unwrap_or_else(|| self.clock.now());
         let new_job = NewJob::builder()
             .id(JobId::new())
             .resident(true)
             .job_type(self.job_type.clone())
             .config(config)?
             .tracing_context(es_entity::context::TracingContext::current())
+            .schedule_at(schedule_at)
             .build()
             .expect("Could not build new job");
-        let mut op = self.repo.begin_op_with_clock(&self.clock).await?;
         match self.repo.create_in_op(&mut op, new_job).await {
-            Ok(mut job) => {
-                let schedule_at = op.maybe_now().unwrap_or_else(|| self.clock.now());
-                // A resident type never wires a `PollerHandle` in -- there is
-                // at most one job of it, ever, so there is no backlog for a
-                // head-swap claim to short-circuit into; this fresh, never-
-                // populated handle just makes `ExecutionInsertHook`'s claim
-                // step a guaranteed no-op, mirroring that exclusion.
+            Ok(job) => {
+                // At most one job of this type ever exists, so there is no
+                // backlog to short-circuit into -- this never-populated
+                // handle makes the claim step a guaranteed no-op.
                 let poller_ref: PollerHandle = Arc::new(std::sync::OnceLock::new());
                 ExecutionInsertHook::register_one(
                     &mut op,
@@ -238,8 +237,6 @@ where
                     },
                 )
                 .await?;
-                job.schedule_execution(schedule_at);
-                self.repo.update_in_op(&mut op, &mut job).await?;
                 op.commit().await?;
                 Ok(self.handle(job.id))
             }
